moved get_public_key_data to gpgmh
